feat: move shared types in a shared package
