+++ apps/api/package.json
@@ -23,6 +23,7 @@
   },
   "dependencies": {
     "@decentdebates/db": "workspace:*",
+    "@decentdebates/shared-types": "workspace:*",
     "@lexical/link": "^0.9.2",
     "@nestjs/common": "9.0.0",
     "@nestjs/config": "^2.2.0",
