+++ apps/api/src/entities/comment/comment.model.ts
@@ -1,19 +1,7 @@
+export { Comment, UpdateCommentData } from '@decentdebates/shared-types';
+
 export interface AddCommentData {
   ticketId: number;
   content: string;
   commenterId: number;
-}
-
-export interface Comment {
-  commentId: number;
-  content: string;
-  commenterId: number;
-  createdAt: string;
-  modifiedAt: string;
-  commenterUsername: string;
-}
-
-export interface UpdateCommentData {
-  commentId: number;
-  content: string;
 }
