+++ apps/api/src/entities/notification/notificatin.model.ts
@@ -1,17 +1,5 @@
-export enum NotificationEvents {
-  ARGUMENT = 'ARGUMENT',
-  DEBATE = 'DEBATE',
-  SUGGESTION = 'SUGGESTION',
-}
-
-export interface Notification {
-  id: number;
-  title: string;
-  content: string;
-  recipientId: number;
-  notificationEvent: NotificationEvents;
-  isRead: boolean;
-}
+export { NotificationEvents, Notification } from '@decentdebates/shared-types';
+import type { Notification } from '@decentdebates/shared-types';
 
 export type NewNotification = Omit<Notification, 'id'>;
