+++ apps/client/package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@decentdebates/shared-types": "workspace:*",
     "@blueprintjs/core": "^4.16.3",
     "@blueprintjs/popover2": "^1.13.3",
     "@blueprintjs/select": "^4.9.12",
