+++ packages/shared-types/src/comment.ts
@@ -1,0 +1,13 @@
+export interface Comment {
+  commentId: number;
+  content: string;
+  commenterId: number;
+  createdAt: string;
+  modifiedAt: string;
+  commenterUsername: string;
+}
+
+export interface UpdateCommentData {
+  commentId: number;
+  content: string;
+}
