+++ packages/shared-types/src/index.ts
@@ -1,0 +1,6 @@
+export * from './comment';
+export * from './debate';
+export * from './moderator';
+export * from './notification';
+export * from './review';
+export * from './user';
