+++ packages/shared-types/src/moderator.ts
@@ -1,0 +1,5 @@
+export enum CardLabels {
+  DEBATE = 'debate',
+  ARGUMENT = 'argument',
+  REPORT = 'report',
+}
