+++ packages/shared-types/src/notification.ts
@@ -1,0 +1,14 @@
+export enum NotificationEvents {
+  ARGUMENT = 'ARGUMENT',
+  DEBATE = 'DEBATE',
+  SUGGESTION = 'SUGGESTION',
+}
+
+export interface Notification {
+  id: number;
+  title: string;
+  content: string;
+  recipientId: number;
+  notificationEvent: NotificationEvents;
+  isRead: boolean;
+}
